restructured in-component services to /services in root
