+++ HRMS.SPA/src/app/common/header/header.component.ts
@@ -1,10 +1,10 @@
 import { NgClass } from '@angular/common';
 import { MatMenuModule } from '@angular/material/menu';
 import { Component, HostListener } from '@angular/core';
-import { ToggleService } from '../sidebar/toggle.service';
+import { ToggleService } from '../../services/toggle.service';
 import { MatButtonModule } from '@angular/material/button';
 import { RouterLink, RouterLinkActive } from '@angular/router';
-import { CustomizerService } from '../../customizer/customizer.service';
+import { CustomizerService } from '../../services/customizer.service';
 
 @Component({
   selector: 'app-header',
