+++ HRMS.SPA/src/app/services/toggle.service.ts
